requirements.txt
django
psycopg2
numpy
gunicorn
django_cleanup
paramiko
psutil
apscheduler
cryptography
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django_reverse_admin